+++ timeline/plugins.py
@@ -48,7 +48,7 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
     icon = 'clock-o'
 
     #: Entry point URL ID (must take project omics_uuid as "project" argument)
-    entry_point_url_id = 'timeline:project_timeline'
+    entry_point_url_id = 'timeline:list_project'
 
     #: Description string
     description = 'Timeline of project events'