+++ timeline/urls.py
@@ -9,13 +9,13 @@ urlpatterns = [
     url(
         regex=r'^(?P<project>[0-9a-f-]+)$',
         view=views.ProjectTimelineView.as_view(),
-        name='project_timeline',
+        name='list_project',
     ),
-    url(    # TODO: Refactor this to use the object ref model uuid?
+    url(
         regex=r'^(?P<project>[0-9a-f-]+)/(?P<object_model>[\w-]+)/'
               r'(?P<object_uuid>[0-9a-f-]+)$',
         view=views.ObjectTimelineView.as_view(),
-        name='object_timeline',
+        name='list_object',
     ),
     # Taskflow API views
     url(